fix(inbox): count a voice batch as one draft everywhere
Several expenses spoken in one breath share a voiceBatchId and fold into
one collapsible inbox row. The waiting-drafts counts did not fold, so one
utterance of four items read as four pending drafts on the dashboard inbox
glyph and the unassigned-captures card.

Extract voiceBatchId into a shared lib alongside foldedCaptureCount, and
route both counts through it so the number matches what the inbox shows.
